News article create / edit fixes
* [Fix] There was a reference to $article->headline in the setup_article
routine where $article was undefined if the article was being created.
This has been fixed so that $article always refers to an article at that
point.
* [Fix] Upon successfully creating / editing an article, the redirection
was attempting to redirect to the action /news/view, which didn't exist
(/news/view_by_id or /news/view_by_url_key do).  Fixed to redirect to
/news/view_by_url_key.
* [Mod] Altered the language file msgids "object.singular.news" and
"object.plural.news" to "object.singular.news-articles" and
"object.plural.news-articles" respectively to match what's referenced in
the event log.  Altered all references to these IDs to match as well.

Fixes #30