Season archive routine
* [New] - Given the ability to archive the current season once all the
matches are either completed or cancelled.  Archiving the seasons means
by definition it is no longer the "current" season and therefore things
that can only be done in the current season can't be done to this season
any more (and can't be done at all until a new season has been created,
which is inherently the current season until it is itself archived).