Bug fixes
* [Fix] Checked that $type is defined in the FilterHTML model before
checking the value.
* [Fix] If a player was being added to a match (and there was no
previous player in that position) but the match had started, the
matches_played was not incremented for the player.  This has been fixed.
* [Fix] When edting or creating a team, the routine that checked for a
player's active membership did not check if they were an active member
of the team being edited.  This has been fixed.
* [Fix] HTML emails erroneously went out with a minimum height style on
the body of 1,000 pixels.  This style atttribute has been removed.

Fixes #40, #41