Bug fix - matches_played, team view page
* [Fix] Individual players matches_played counts weren't updated if
there was a player position with no player (player missing).  This is
because the matches_played are incremented when the match_started flag
is set to 1, which was controlled by the presence of games with the
complete flag set to 1, however when a player was missing, the games
they were playing were automatically marked as complete already, so when
the first game was updated. the match was already seen as being
complete.  The check to see if the match is started now checks for games
that have been started *and* completed, so it doesn't take into account
awarded games that were never started.
* [Fix] The team view page was incorrectly linking the team's division
name to the captain's person - this has been fixed to link to the league
table.

Closes #53, closes #54